+++ crates/autopilot/src/domain/competition/participation_guard/db.rs
@@ -1,0 +1,104 @@
+use {
+    crate::{
+        database::Postgres,
+        domain::{eth, Metrics},
+    },
+    ethrpc::block_stream::CurrentBlockWatcher,
+    std::{
+        sync::Arc,
+        time::{Duration, Instant},
+    },
+};
+
+/// Checks the DB by searching for solvers that won N last consecutive auctions
+/// but never settled any of them.
+#[derive(Clone)]
+pub(super) struct Validator(Arc<Inner>);
+
+struct Inner {
+    db: Postgres,
+    banned_solvers: dashmap::DashMap<eth::Address, Instant>,
+    ttl: Duration,
+    last_auctions_count: u32,
+}
+
+impl Validator {
+    pub fn new(
+        db: Postgres,
+        current_block: CurrentBlockWatcher,
+        settlement_updates_receiver: tokio::sync::mpsc::UnboundedReceiver<()>,
+        ttl: Duration,
+        last_auctions_count: u32,
+    ) -> Self {
+        let self_ = Self(Arc::new(Inner {
+            db,
+            banned_solvers: Default::default(),
+            ttl,
+            last_auctions_count,
+        }));
+
+        self_.start_maintenance(settlement_updates_receiver, current_block);
+
+        self_
+    }
+
+    /// Update the internal cache only once the settlement table is updated to
+    /// avoid redundant DB queries.
+    fn start_maintenance(
+        &self,
+        mut settlement_updates_receiver: tokio::sync::mpsc::UnboundedReceiver<()>,
+        current_block: CurrentBlockWatcher,
+    ) {
+        let self_ = self.0.clone();
+        tokio::spawn(async move {
+            while settlement_updates_receiver.recv().await.is_some() {
+                let current_block = current_block.borrow().number;
+                match self_
+                    .db
+                    .find_non_settling_solvers(self_.last_auctions_count, current_block)
+                    .await
+                {
+                    Ok(non_settling_solvers) => {
+                        let non_settling_solvers = non_settling_solvers
+                            .into_iter()
+                            .map(|solver| {
+                                let address = eth::Address(solver.0.into());
+
+                                Metrics::get()
+                                    .non_settling_solver
+                                    .with_label_values(&[&format!("{:#x}", address.0)]);
+
+                                address
+                            })
+                            .collect::<Vec<_>>();
+
+                        tracing::debug!(?non_settling_solvers, "found non-settling solvers");
+
+                        let now = Instant::now();
+                        for solver in non_settling_solvers {
+                            self_.banned_solvers.insert(solver, now);
+                        }
+                    }
+                    Err(err) => {
+                        tracing::warn!(?err, "error while searching for non-settling solvers")
+                    }
+                }
+            }
+        });
+    }
+}
+
+#[async_trait::async_trait]
+impl super::Validator for Validator {
+    async fn is_allowed(&self, solver: &eth::Address) -> anyhow::Result<bool> {
+        if let Some(entry) = self.0.banned_solvers.get(solver) {
+            if Instant::now().duration_since(*entry.value()) < self.0.ttl {
+                return Ok(false);
+            } else {
+                self.0.banned_solvers.remove(solver);
+            }
+        }
+
+        Ok(true)
+    }
+}
