+++ crates/autopilot/src/domain/competition/participation_guard/onchain.rs
@@ -1,0 +1,20 @@
+use crate::{domain::eth, infra::Ethereum};
+
+/// Calls Authenticator contract to check if a solver has a sufficient
+/// permission.
+pub(super) struct Validator {
+    pub eth: Ethereum,
+}
+
+#[async_trait::async_trait]
+impl super::Validator for Validator {
+    async fn is_allowed(&self, solver: &eth::Address) -> anyhow::Result<bool> {
+        Ok(self
+            .eth
+            .contracts()
+            .authenticator()
+            .is_solver(solver.0)
+            .call()
+            .await?)
+    }
+}
